Migrated all record components to functional components. Added rjsf types.
