+++ src/components/record/RecordCreate.tsx
@@ -1,3 +1,5 @@
+import React, { useCallback } from "react";
+
 import type {
   SessionState,
   BucketState,
@@ -5,8 +7,6 @@ import type {
   CollectionRouteMatch,
   Capabilities,
 } from "../../types";
-
-import React, { PureComponent } from "react";
 
 import * as CollectionActions from "../../actions/collection";
 import RecordForm from "./RecordForm";
@@ -27,42 +27,49 @@ export type Props = OwnProps &
     createRecord: typeof CollectionActions.createRecord;
   };
 
-export default class RecordCreate extends PureComponent<Props> {
-  onSubmit = ({ __attachment__: attachment, ...record }: any) => {
-    const { match, createRecord } = this.props;
-    const {
-      params: { bid, cid },
-    } = match;
-    createRecord(bid, cid, record, attachment);
-  };
+export default function RecordCreate({
+  match,
+  session,
+  capabilities,
+  bucket,
+  collection,
+  createRecord,
+}: Props) {
+  const onSubmit = useCallback(
+    ({ __attachment__: attachment, ...record }) => {
+      const {
+        params: { bid, cid },
+      } = match;
+      createRecord(bid, cid, record, attachment);
+    },
+    [match, createRecord]
+  );
 
-  render() {
-    const { match, session, bucket, collection, capabilities } = this.props;
-    const {
-      params: { bid, cid },
-    } = match;
-    return (
-      <div>
-        <h1>
-          Add a new record in{" "}
-          <b>
-            {bid}/{cid}
-          </b>
-        </h1>
-        <div className="card">
-          <div className="card-body">
-            <RecordForm
-              bid={bid}
-              cid={cid}
-              session={session}
-              bucket={bucket}
-              collection={collection}
-              onSubmit={this.onSubmit}
-              capabilities={capabilities}
-            />
-          </div>
+  const {
+    params: { bid, cid },
+  } = match;
+
+  return (
+    <div>
+      <h1>
+        Add a new record in{" "}
+        <b>
+          {bid}/{cid}
+        </b>
+      </h1>
+      <div className="card">
+        <div className="card-body">
+          <RecordForm
+            bid={bid}
+            cid={cid}
+            session={session}
+            bucket={bucket}
+            collection={collection}
+            onSubmit={onSubmit}
+            capabilities={capabilities}
+          />
         </div>
       </div>
-    );
-  }
+    </div>
+  );
 }
